backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
faiss-cpu
mistralai
langchain-text-splitters
jinja2
